Update type aware search doc to use recipe-examples
Previously, the doc was using examples that I found in random
repositories. However, as people change code over time, that could
mean that the tutorials go out of date. I created a recipe-examples
repository to store code in that I can use to highlight examples
in these tutorials.

On top of that, I switched the example from `createFile` to
`createTempFile` to lead in to another tutorial where we'll use
a security recipe to update the file.