Rewrite agent documentation
This is a substantial change to the approach we've had with agent docs.
Instead of have a folder that has a doc for each possible step and
relying on people to click on the ones that apply to them, this creates
a base "Agent Configuration" doc that provides _all_ of the steps needed
to configure the agent. It then links out to the specifics for each step
to the existing docs.

Almost every doc has been modified to be clearer and to reduce inconsistencies
between them. A new doc has also been added that shows all possible agent
configuration variables.

To assist in the review of this, I've copied over the files (not the images)
to our Moderne knowledge base for easier review: https://app.gitbook.com/o/-MEp_3EtccewzekKY8mZ/s/JC9dRbwVINQjAyoDyBuW/agent-configuration