Update CLI example
I found the current explanation to be somewhat confusing. It wasn't
clear to me what benefits you'd get from running a recipe on a
remote LST or that you couldn't run against multiple repositories
locally.

I've attempted to rephrase things to be a bit clearer and flow
a bit better.