Update LST artifact doc
Right now, the doc refers to only the Moderne build plugins.
Furthermore, some of the information on those plugins are incorrect.

This updates the doc to include the Moderne CLI, new images, and
fixes for all parts of the doc.

This partially addresses this issue: https://github.com/moderneinc/moderne-cli/issues/527