Update recipe-authoring.md
As you can't run test with Java 23¹, it's better to just define the Java version as 17 or 21 in the documentation (and maybe 25 once it's released in September next year). The user reading the documentation does not want to understand the nitty gritty things about OpenRewrite, but wants to run the tests. 

 __
¹ You will run into a "Failed to parse sources or run recipe => Failed to convert for the following cursor stack" error.