Update configuring CLI for DX to be more generic
Olga said that most of these instructions are ones that would apply to other types of customers (such as Enterprise customers). Because of that, I've updated the doc to remove DX specific language in favor of talking about internal tools. I also updated the getting started CLI guide to point to this.